Move data loading into a common module
This will allow at least an intermediate stage where the C
implementation calls into Python to load the data from the file-like
object and populates itself from the results.
