+++ src/zoneinfo/_zoneinfo.py
@@ -9,7 +9,7 @@ import sys
 import weakref
 from datetime import datetime, timedelta, timezone, tzinfo
 
-from . import _tzpath
+from . import _common, _tzpath
 
 EPOCH = datetime(1970, 1, 1)
 EPOCHORDINAL = datetime(1970, 1, 1).toordinal()
@@ -217,7 +217,7 @@ class ZoneInfo(tzinfo):
 
     def _load_file(self, fobj):
         # Retrieve all the data as it exists in the zoneinfo file
-        trans_idx, trans_utc, utcoff, isdst, abbr, tz_str = self._load_data(
+        trans_idx, trans_utc, utcoff, isdst, abbr, tz_str = _common.load_data(
             fobj
         )
 
@@ -255,96 +255,6 @@ class ZoneInfo(tzinfo):
             self._tz_after = _parse_tz_str(tz_str)
         else:
             self._tz_after = self._ttinfos[-1]
-
-    @classmethod
-    def _load_data(cls, fobj):
-        header = _TZifHeader.from_file(fobj)
-
-        if header.version == 1:
-            time_size = 4
-            time_type = "l"
-        else:
-            # Version 2+ has 64-bit integer transition times
-            time_size = 8
-            time_type = "q"
-
-            # Version 2+ also starts with a Version 1 header and data, which
-            # we need to skip now
-            skip_bytes = (
-                header.timecnt * 5
-                + header.typecnt * 6  # Transition times and types
-                + header.charcnt  # Local time type records
-                + header.leapcnt * 8  # Time zone designations
-                + header.isstdcnt  # Leap second records
-                + header.isutcnt  # Standard/wall indicators  # UT/local indicators
-            )
-
-            fobj.seek(skip_bytes, 1)
-
-            # Now we need to read the second header, which is not the same
-            # as the first
-            header = _TZifHeader.from_file(fobj)
-
-        typecnt = header.typecnt
-        timecnt = header.timecnt
-        charcnt = header.charcnt
-
-        # The data portion starts with timecnt transitions and indices
-        if timecnt:
-            trans_list_utc = struct.unpack(
-                f">{timecnt}{time_type}", fobj.read(timecnt * time_size)
-            )
-            trans_idx = struct.unpack(f">{timecnt}B", fobj.read(timecnt))
-        else:
-            trans_list_utc = []
-            trans_idx = []
-
-        # Read the ttinfo struct, (utoff, isdst, abbrind)
-        if typecnt:
-            utcoff, isdst, abbrind = zip(
-                *(struct.unpack(">lbb", fobj.read(6)) for i in range(typecnt))
-            )
-        else:
-            utcoff = ()
-            isdst = ()
-            abbrind = ()
-
-        # Now read the abbreviations. They are null-terminated strings, indexed
-        # not by position in the array but by position in the unsplit
-        # abbreviation string. I suppose this makes more sense in C, which uses
-        # null to terminate the strings, but it's inconvenient here...
-        char_total = 0
-        abbr_vals = {}
-        for abbr in fobj.read(charcnt).decode().split("\x00"):
-            abbr_vals[char_total] = abbr
-            char_total += len(abbr) + 1
-
-        abbr = [abbr_vals[idx] for idx in abbrind]
-
-        # The remainder of the file consists of leap seconds (currently unused) and
-        # the standard/wall and ut/local indicators, which are metadata we don't need.
-        # In version 2 files, we need to skip the unnecessary data to get at the TZ string:
-        if header.version >= 2:
-            # Each leap second record has size (time_size * 4)
-            skip_bytes = header.isutcnt + header.isstdcnt + header.leapcnt * 32
-            fobj.seek(skip_bytes, 1)
-
-            c = fobj.read(1)  # Should be \n
-            assert c == b"\n", c
-
-            tz_bytes = b""
-            # TODO: Walrus operator
-            while True:
-                c = fobj.read(1)
-                if c == b"\n":
-                    break
-                tz_bytes += c
-
-            tz_str = tz_bytes.decode()
-        else:
-            tz_str = None
-
-        return trans_idx, trans_list_utc, utcoff, isdst, abbr, tz_str
 
     @staticmethod
     def _utcoff_to_dstoff(trans_idx, utcoffsets, isdsts):
@@ -442,39 +352,6 @@ class ZoneInfo(tzinfo):
             trans_list_wall[1][i] += offset_1
 
         return trans_list_wall
-
-
-class _TZifHeader:
-    __slots__ = [
-        "version",
-        "isutcnt",
-        "isstdcnt",
-        "leapcnt",
-        "timecnt",
-        "typecnt",
-        "charcnt",
-    ]
-
-    def __init__(self, *args):
-        assert len(self.__slots__) == len(args)
-        for attr, val in zip(self.__slots__, args):
-            setattr(self, attr, val)
-
-    @classmethod
-    def from_file(cls, stream):
-        # The header starts with a 4-byte "magic" value
-        if stream.read(4) != b"TZif":
-            raise ValueError("Invalid TZif file: magic not found")
-
-        version = int(stream.read(1))
-        stream.read(15)
-
-        args = (version,)
-
-        # Slots are defined in the order that the bytes are arranged
-        args = args + struct.unpack(">6l", stream.read(24))
-
-        return cls(*args)
 
 
 class _ttinfo:
